+++ examples/box1L/generate_box1L.py
@@ -32,11 +32,4 @@ if __name__ == "__main__":
 
         # the highest order of the final epsilon expansion --> change this value to whatever you think is appropriate
         requested_orders = [0],
-
-        # the optimization level to use in FORM (can be 0, 1, 2, 3, 4)
-        form_optimization_level = 2,
-
-        # the method to be used for the sector decomposition
-        # valid values are ``iterative`` or ``geometric`` or ``geometric_ku``
-        decomposition_method = 'iterative',
     )