+++ examples/elliptic2L_euclidean/generate_elliptic2L_euclidean.py
@@ -41,7 +41,7 @@ if __name__ == "__main__":
 
     # the method to be used for the sector decomposition
     # valid values are ``iterative`` and ``geometric``
-    decomposition_method = 'iterative',
+    decomposition_method = 'geometric',
 
     contour_deformation = False