+++ examples/elliptic2L_physical/generate_elliptic2L_physical.py
@@ -42,7 +42,7 @@ if __name__ == "__main__":
 
     # the method to be used for the sector decomposition
     # valid values are ``iterative`` and ``geometric``
-    decomposition_method = 'iterative',
+    decomposition_method = 'geometric',
 
     contour_deformation = True,