+++ examples/triangle3L/generate_triangle3L.py
@@ -46,7 +46,7 @@ if __name__ == "__main__":
 
     # the method to be used for the sector decomposition
     # valid values are ``iterative`` or ``geometric`` or ``geometric_ku``
-    decomposition_method = 'iterative',
+    decomposition_method = 'geometric',
 
     contour_deformation = False